requirements.txt
attr==0.3.2
blinker==1.7.0
bpython==0.24
brotli==1.1.0
brotlicffi==1.1.0.0
cffi==1.16.0
ConfigParser==6.0.0
contextlib2==21.6.0
cx_Oracle==8.3.0
Cython==3.0.6
dl==0.1.0
docutils==0.20.1
filelock==3.13.1
HTMLParser==0.0.2
httplib2==0.22.0
httpretty==1.1.4
ipdb==0.13.13
ipython==8.18.1
ipywidgets==8.1.1
jnius==1.1.0
jose==1.0.0
keyring==24.3.0
mock==5.1.0
MySQL-python==1.2.5
numpy==1.26.2
onelogin==3.1.6
Pillow==10.1.0
psycopg==3.1.14
psycopg2==2.9.9
psycopg2cffi==2.9.0
pycryptodome==3.19.0
pycurl==7.45.2
pylibmc==1.6.3
pymemcache==4.0.0
pyOpenSSL==23.3.0
pytest==7.4.3
pytest_django==4.7.0
python_bcrypt==0.3.2
pytz==2023.3.post1
pywatchman==1.4.1
PyYAML==6.0.1
qrcode==7.4.2
railroad==0.5.0
simplejson==3.19.2
Sphinx==7.2.6
tblib==3.0.0
thread==0.1.2
tornado==6.4
trove_classifiers==2023.11.29
urllib3_secure_extra==0.1.0
xmlrpclib==1.0.1
zstandard==0.22.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-crispy-forms==2.1
docopt==0.6.2
idna==3.6
oauthlib==3.2.2
pipreqs==0.4.13
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9